+++ wasm-tests/run-all.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -e
+
+./webtransport-tests/run.sh
